requirements.txt
streamlit
numpy
pandas
pip
scikit-learn
Pillow
